Fix possible truncation when calculating hashsize().
In Windows, long is only 32-bits, so the shift operation in hashsize() could
result in a bogus calculation for very large hash tables. This commit fixes that
by ensuring the shift operation operates on a size_t.

This addresses the following complaint from MSVC:

1>cuckoohash_map.hh(895): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)

+++ src/cuckoohash_map.hh
@@ -892,7 +892,7 @@ private:
     // hashsize returns the number of buckets corresponding to a given
     // hashpower.
     static inline size_t hashsize(const size_t hp) {
-        return 1UL << hp;
+        return size_t(1) << hp;
     }
 
     // hashmask returns the bitmask for the buckets array corresponding to a
